Moved driver from drivers/ssd1680 to drivers/epaper and removed useless demo
